requirements.txt
matplotlib       == 3.7.*
numpy            == 1.24.*
numpy-quaternion == 2022.*
